+++ Project-1/src/com/samvbeckmann/ai/project1/Runner.java
@@ -11,7 +11,7 @@ import java.util.Scanner;
 public class Runner
 {
     private static final int NUM_TESTS = 1000;
-    private static final float VALUE_EPSILON = 0.001F;
+    private static final double VALUE_EPSILON = 0.001;
 
     public static void main(String[] args)
     {
@@ -35,25 +35,25 @@ public class Runner
                 case "exampleMaze":
                     if (mazeParams.length != 3)
                         throw new InvalidFileFormatException("Maze argument list not correct length!");
-                    maze = new ExampleMaze(Float.parseFloat(mazeParams[0]),
-                            Float.parseFloat(mazeParams[1]),
-                            Float.parseFloat(mazeParams[2]));
+                    maze = new ExampleMaze(Double.parseDouble(mazeParams[0]),
+                            Double.parseDouble(mazeParams[1]),
+                            Double.parseDouble(mazeParams[2]));
                     break;
                 case "wumpusWorld":
                     if (mazeParams.length != 6)
                         throw new InvalidFileFormatException("Maze argument list not correct length!");
-                    maze = new WumpusWorld(Float.parseFloat(mazeParams[0]),
-                            Float.parseFloat(mazeParams[1]),
-                            Float.parseFloat(mazeParams[2]),
-                            Float.parseFloat(mazeParams[3]),
-                            Float.parseFloat(mazeParams[4]),
-                            Float.parseFloat(mazeParams[5]));
+                    maze = new WumpusWorld(Double.parseDouble(mazeParams[0]),
+                            Double.parseDouble(mazeParams[1]),
+                            Double.parseDouble(mazeParams[2]),
+                            Double.parseDouble(mazeParams[3]),
+                            Double.parseDouble(mazeParams[4]),
+                            Double.parseDouble(mazeParams[5]));
                     break;
                 default:
                     throw new InvalidFileFormatException("Maze type not supported!");
             }
 
-            Float discount = Float.parseFloat(br.readLine());
+            double discount = Double.parseDouble(br.readLine());
 
             policyValueIteration = constructAndReportPolicy(new StrategyValueIteration(discount, VALUE_EPSILON), maze);
             policyPolicyIteration = constructAndReportPolicy(new StrategyPolicyIteration(discount), maze);
@@ -86,8 +86,8 @@ public class Runner
         System.out.println("Value Iteration Policy:\n\n" + valueIterationViz);
         System.out.println("Policy Iteration Policy:\n\n" + policyIterationViz);
 
-        float totalRewardPolicy = 0;
-        float totalRewardValue = 0;
+        double totalRewardPolicy = 0;
+        double totalRewardValue = 0;
         for (int i = 0; i < NUM_TESTS; i++)
         {
             Simulation policySimulation = new Simulation(maze.getStartingState(), policyPolicyIteration);
@@ -96,8 +96,8 @@ public class Runner
             totalRewardValue += valueSimulation.performSimulation();
         }
 
-        float netRewardPolicy = totalRewardPolicy / NUM_TESTS;
-        float netRewardValue = totalRewardValue / NUM_TESTS;
+        double netRewardPolicy = totalRewardPolicy / NUM_TESTS;
+        double netRewardValue = totalRewardValue / NUM_TESTS;
 
         System.out.println("Average reward for Policy Iteration: " + netRewardPolicy);
         System.out.println("Average reward for Value Iteration:  " + netRewardValue);
