+++ Project-1/src/com/samvbeckmann/ai/project1/State.java
@@ -12,7 +12,7 @@ import java.util.Map;
  */
 public class State
 {
-    private float reward;
+    private double reward;
     private TransitionMap transitions;
     private boolean isTerminal;
     private boolean isInitialized;
@@ -27,7 +27,7 @@ public class State
         isInitialized = true;
     }
 
-    public State(String id, float reward, boolean isTerminal)
+    public State(String id, double reward, boolean isTerminal)
     {
         this.id = id;
         this.reward = reward;
@@ -46,7 +46,7 @@ public class State
         return isTerminal;
     }
 
-    public float getReward()
+    public double getReward()
     {
         return reward;
     }
@@ -57,9 +57,9 @@ public class State
      * @param action Action that would be taken.
      * @return Transition Table as a result of that action.
      */
-    public Map<State, Float> getPossibleOutcomes(Action action)
+    public Map<State, Double> getPossibleOutcomes(Action action)
     {
-        Map<State, Float> transitionTable;
+        Map<State, Double> transitionTable;
         try
         {
             if (!isInitialized)
